I fixed a bug where I had the time stamps recording as floats instead of longs. This caused the the MS since the UNIX epoch based time stamp to be recorded as #####E## or some such value instead of the full true value. The side effect of this was that the times were not changing properly.

+++ pTracker/src/edu/temple/ptracker/activityData.java
@@ -23,11 +23,11 @@ public class ActivityData {
 	}
 		
 	//Add X,Y,Z data along with Time to the temp array.
-	public void addXYZData(float timeData, float xData, float yData, float zData, String name){
+	public void addXYZData(long timeData, float xData, float yData, float zData, String name){
 		
 		//Combine all the inputs into a string to add to the list.
 		String tempString = new String();
-		tempString = Float.toString(timeData) + "," + Float.toString(xData) + "," + Float.toString(yData) + "," + Float.toString(zData) + "," + name;
+		tempString = Long.toString(timeData) + "," + Float.toString(xData) + "," + Float.toString(yData) + "," + Float.toString(zData) + "," + name;
 		activityVector.add(tempString);
 		
 	}
